Cardano.Api.UTxO

Description

NOTE (1): This module is meant to be imported qualified as UTxO.

NOTE (2): This module is name-spaces slightly different from the rest because it is meant to be used as a replacement of the UTxO type of the cardano-api which is not convenient enough to work with. Having it as UTxO causes cyclic imports with other modules also relying on this newtype. So instead, we do 'as if' it was part of the cardano-api in the first place.

Documentation

type Era = ConwayEra

type UTxO = UTxO' (TxOut CtxUTxO Era)

newtype UTxO' out

Newtype with phantom types mostly required to work around the poor interface of UTXO and provide Monoid and Foldable instances to make utxo manipulation bareable.

Constructors

UTxO

Fields

toMap :: Map TxIn out

fromList :: [(TxIn, out)] -> UTxO' out

Create a UTxO from a list of TxIn and out pairs.

singleton :: TxIn -> out -> UTxO' out

Create a UTxO from a single unspent transaction output.

resolveTxIn :: TxIn -> UTxO' out -> Maybe out

Find an out for a given TxIn.

toList :: UTxO' out -> [(TxIn, out)]

Turn a UTxO into a list of pairs.

find :: (out -> Bool) -> UTxO' out -> Maybe (TxIn, out)

Find first UTxO using the output in predicate.

findBy :: ((TxIn, out) -> Bool) -> UTxO' out -> Maybe (TxIn, out)

Find first UTxO using both input and output in predicate.

filter :: (out -> Bool) -> UTxO' out -> UTxO' out

Filter UTxO to only include outs satisfying given predicate.

inputSet :: UTxO' out -> Set TxIn

Get the UTxO domain input's set

render :: (TxIn, TxOut ctx era) -> Text

Get a human-readable pretty text representation of a UTxO.

difference :: UTxO' out -> UTxO' out -> UTxO' out

Remove the right hand side from the left hand side.

containsOutputs :: UTxO -> UTxO -> Bool

Check if the first UTxO contains all **outputs** of the second UTxO. First argument is the UTxO to search in, second argument is the UTxO to search for.

map :: (TxOut CtxUTxO Era -> TxOut CtxUTxO Era) -> UTxO -> UTxO

foldMap :: Monoid m => (TxOut CtxUTxO Era -> m) -> UTxO -> m

txOutputs :: UTxO -> [TxOut CtxUTxO Era]

null :: UTxO -> Bool

size :: UTxO -> Int

totalValue :: UTxO -> Value

totalLovelace :: UTxO -> Coin

Type Conversions

fromApi :: UTxO era -> UTxO

Transforms a UTxO containing tx outs from any era into Babbage era.

toApi :: UTxO -> UTxO Era