Hydra.Cardano.Api.ChainPoint

Documentation

getChainPoint :: BlockHeader -> ChainPoint

Get the chain point corresponding to a given BlockHeader.

Generators

genChainPoint :: Gen ChainPoint

Generate a chain point with a likely invalid block header hash.

genChainPointAt :: SlotNo -> Gen ChainPoint

Generate a chain point at given slot with a likely invalid block header hash.

Orphans

Orphan instances

Arbitrary ChainPoint

Instance details

Methods

arbitrary :: Gen ChainPoint

shrink :: ChainPoint -> [ChainPoint]