Hydra.Cardano.Api.Point

Documentation

toConsensusPointHF :: HeaderHash block ~ OneEraHash xs => ChainPoint -> Point block

TODO: This function is available in more recent versions of the cardano-api, once we upgrade the dependency set, we can ditch it.

fromConsensusPointHF :: HeaderHash block ~ OneEraHash xs => Point block -> ChainPoint

Orphan instances

Ord ChainPoint

Instance details

Methods

compare :: ChainPoint -> ChainPoint -> Ordering

(<) :: ChainPoint -> ChainPoint -> Bool

(<=) :: ChainPoint -> ChainPoint -> Bool

(>) :: ChainPoint -> ChainPoint -> Bool

(>=) :: ChainPoint -> ChainPoint -> Bool

max :: ChainPoint -> ChainPoint -> ChainPoint

min :: ChainPoint -> ChainPoint -> ChainPoint