Hydra.Cardano.Api.ScriptData

Extras

type ToScriptData a = ToData a

Data-types that can be marshalled into a generic ScriptData structure.

type FromScriptData a = FromData a

Data-types that can be unmarshalled from a generic ScriptData structure.

toScriptData :: ToScriptData a => a -> HashableScriptData

Serialise some type into a generic script data.

fromScriptData :: FromScriptData a => HashableScriptData -> Maybe a

Deserialise some generic script data into some type.

txOutScriptData :: TxOut CtxTx era -> Maybe HashableScriptData

Get the HashableScriptData associated to the a TxOut. Note that this requires the CtxTx context.

Type Conversions

fromLedgerData :: Data era -> HashableScriptData

Convert a cardano-ledger script Data into a cardano-api ScriptDatum.

toLedgerData :: Era era => HashableScriptData -> Data era

Convert a cardano-api script data into a cardano-ledger script Data. XXX: This is a partial function. Ideally it would fall back to the Data portion in HashableScriptData.