Hydra.Cardano.Api.Tx

Extras

signTx :: IsShelleyBasedEra era => SigningKey PaymentKey -> Tx era -> Tx era

Sign transaction using the provided secret key It only works for tx not containing scripts. You can't sign a script utxo with this.

utxoProducedByTx :: Tx Era -> UTxO

Get the UTxO that are produced by some transaction. XXX: Defined here to avoid cyclic module dependency

txFee' :: HasCallStack => Tx era -> Lovelace

Get explicit fees allocated to a transaction.

NOTE: this function is partial and throws if given a Byron transaction for which fees are necessarily implicit.

totalExecutionCost :: PParams (ShelleyLedgerEra Era) -> Tx Era -> Lovelace

Calculate the total execution cost of a transaction, according to the budget assigned to each redeemer.

Type Conversions

toLedgerTx :: Tx Era -> Tx (ShelleyLedgerEra Era)

Convert a cardano-api Tx into a cardano-ledger Tx in the Babbage era

fromLedgerTx :: Tx (ShelleyLedgerEra Era) -> Tx Era

Convert a cardano-ledger's Tx in the Babbage era into a cardano-api Tx.