Hydra.Cardano.Api.TxIn

Extras

mkTxIn :: Tx era -> Word -> TxIn

Create a TxIn (a.k.a UTXO) from a transaction and output index.

txIns' :: Tx era -> [TxIn]

Access inputs of a transaction, as an ordered list.

txInputSet :: Tx era -> Set TxIn

Access inputs of a transaction, as an ordered set.

Type Conversions

fromLedgerTxIn :: TxIn StandardCrypto -> TxIn

Convert a cardano-ledger TxIn into a cardano-api TxIn

toLedgerTxIn :: TxIn -> TxIn StandardCrypto

Convert a cardano-api TxIn into a cardano-ledger TxIn

fromPlutusTxOutRef :: TxOutRef -> TxIn

Convert a plutus' TxOutRef into a cardano-api TxIn

toPlutusTxOutRef :: TxIn -> TxOutRef

Convert a cardano-api TxIn into a plutus TxOutRef.