Hydra.Cardano.Api.Witness

Documentation

mkScriptWitness :: forall ctx era lang. (IsPlutusScriptLanguage lang, HasScriptLanguageInEra lang era) => PlutusScript lang -> ScriptDatum ctx -> ScriptRedeemer -> ScriptWitness ctx era

Construct a full script witness from a datum, a redeemer and a full Script. That witness has no execution budget.

mkScriptReference :: forall ctx era lang. (IsPlutusScriptLanguage lang, HasScriptLanguageInEra lang era) => TxIn -> PlutusScript lang -> ScriptDatum ctx -> ScriptRedeemer -> ScriptWitness ctx era

Construct a reference script witness, only referring to a TxIn which is expected to contain the given script (only required to satisfy types).