Cardano.Api.UTxO

Description

NOTE (1): This module is meant to be imported qualified as UTxO.

NOTE (2): This module is name-spaces slightly different from the rest because it is meant to be used as a replacement of the UTxO type of the cardano-api which is not convenient enough to work with. Having it as UTxO causes cyclic imports with other modules also relying on this newtype. So instead, we do 'as if' it was part of the cardano-api in the first palce.

Documentation

type Era = BabbageEra

type UTxO = UTxO' (TxOut CtxUTxO Era)

newtype UTxO' out

Newtype with phantom types mostly required to work around the poor interface of UTXOand provide Monoid and Foldable instances to make utxo manipulation bareable.

Constructors

UTxO

Fields

toMap :: Map TxIn out

fromPairs :: [(TxIn, out)] -> UTxO' out

Create a UTxO from a list of TxIn and out pairs.

singleton :: (TxIn, TxOut CtxUTxO Era) -> UTxO

Create a UTxO from a single unspent transaction output.

resolve :: TxIn -> UTxO' out -> Maybe out

Find an out for a given TxIn.

pairs :: UTxO' out -> [(TxIn, out)]

Turn a UTxO into a list of pairs.

filter :: (out -> Bool) -> UTxO' out -> UTxO' out

Filter UTxO to only include outs satisfying given predicate.

inputSet :: UTxO' out -> Set TxIn

Get the UTxO domain input's set

render :: (TxIn, TxOut ctx era) -> Text

Get a human-readable pretty text representation of a UTxO.

min :: UTxO -> UTxO

Select the minimum (by TxIn) utxo entry from the UTxO map.

This function is partial.

Type Conversions

fromApi :: UTxO Era -> UTxO

toApi :: UTxO -> UTxO Era