Hydra.Cardano.Api.PlutusScript

Type Conversions

fromLedgerScript :: HasCallStack => Script era -> PlutusScript lang

Convert a cardano-ledger Script into a cardano-api PlutusScript

Note that this function is unsafe in two manners:

(a) If the given script is a timelock script, it throws an impure exception; (b) If the given script is in a wrong language, it silently coerces it.

toLedgerScript :: forall lang. HasPlutusScriptVersion lang => PlutusScript lang -> Script (ShelleyLedgerEra Era)

Convert a cardano-api PlutusScript into a cardano-ledger Script.

fromPlutusScript :: Script -> PlutusScript lang

Convert a plutus Script into a cardano-api PlutusScript