Hydra.Cardano.Api.ScriptData

Extras

type ToScriptData a = ToData a

Data-types that can be marshalled into a generic ScriptData structure.

type FromScriptData a = FromData a

Data-types that can be unmarshalled from a generic ScriptData structure.

toScriptData :: ToScriptData a => a -> HashableScriptData

Serialise some type into a generic script data.

fromScriptData :: FromScriptData a => HashableScriptData -> Maybe a

Deserialise some generic script data into some type.

txOutScriptData :: TxOut CtxTx era -> Maybe HashableScriptData

Get the HashableScriptData associated to the a TxOut. Note that this requires the CtxTx context. To get script data in a CtxUTxO context, see lookupScriptData.

lookupScriptData :: forall era. (UsesStandardCrypto era, Era (ShelleyLedgerEra era)) => Tx era -> TxOut CtxUTxO era -> Maybe HashableScriptData

Lookup included datum of given TxOut.

Type Conversions

fromLedgerData :: Data era -> HashableScriptData

Convert a cardano-ledger script Data into a cardano-api ScriptDatum.

toLedgerData :: Era era => HashableScriptData -> Data era

Convert a cardano-api script data into a cardano-ledger script Data. XXX: This is a partial function. Ideally it would fall back to the Data portion in HashableScriptData.

Orphans

Orphan instances

Arbitrary HashableScriptData

Instance details

Methods

arbitrary :: Gen HashableScriptData

shrink :: HashableScriptData -> [HashableScriptData]

Arbitrary ScriptData

Instance details

Methods

arbitrary :: Gen ScriptData

shrink :: ScriptData -> [ScriptData]

FromJSON HashableScriptData

Instance details

Methods

parseJSON :: Value -> Parser HashableScriptData

parseJSONList :: Value -> Parser [HashableScriptData]

omittedField :: Maybe HashableScriptData

FromJSON ScriptData

Instance details

Methods

parseJSON :: Value -> Parser ScriptData

parseJSONList :: Value -> Parser [ScriptData]

omittedField :: Maybe ScriptData

ToJSON HashableScriptData

Instance details

Methods

toJSON :: HashableScriptData -> Value

toEncoding :: HashableScriptData -> Encoding

toJSONList :: [HashableScriptData] -> Value

toEncodingList :: [HashableScriptData] -> Encoding

omitField :: HashableScriptData -> Bool

ToJSON ScriptData

Instance details

Methods

toJSON :: ScriptData -> Value

toEncoding :: ScriptData -> Encoding

toJSONList :: [ScriptData] -> Value

toEncodingList :: [ScriptData] -> Encoding

omitField :: ScriptData -> Bool