Hydra.Cardano.Api.ScriptData

Extras

type ToScriptData a = ToData a

Data-types that can be marshalled into a generic ScriptData structure.

type FromScriptData a = FromData a

Data-types that can be unmarshalled from a generic ScriptData structure.

toScriptData :: ToScriptData a => a -> ScriptData

Serialise some type into a generic ScriptData structure.

getScriptData :: TxOut CtxTx era -> Maybe ScriptData

Get the ScriptData associated to the a TxOut. Note that this requires the CtxTx context. To get script data in a CtxUTxO context, see lookupScriptData.

lookupScriptData :: forall era. (UsesStandardCrypto era, Typeable (ShelleyLedgerEra era)) => Tx era -> TxOut CtxUTxO era -> Maybe ScriptData

Lookup included datum of given TxOut.

Type Conversions

fromLedgerData :: Data era -> ScriptData

Convert a cardano-ledger script Data into a cardano-api ScriptDatum.

toLedgerData :: ScriptData -> Data era

Convert a cardano-api ScriptData into a cardano-ledger script Data.