Hydra.Cardano.Api.Tx

Extras

signTx :: IsShelleyBasedEra era => SigningKey PaymentKey -> Tx era -> Tx era

Sign transaction using the provided secret key It only works for tx not containing scripts. You can't sign a script utxo with this.

txSpendingUTxO :: UTxO -> Tx Era

Create a transaction spending all given UTxO.

fromLedgerTx :: IsShelleyBasedEra era => Tx (ShelleyLedgerEra era) -> Tx era

Convert a cardano-ledger's Tx in the Babbage era into a cardano-api Tx.

utxoProducedByTx :: Tx Era -> UTxO

Get the UTxO that are produced by some transaction. XXX: Defined here to avoid cyclic module dependency

toLedgerTx :: Tx era -> Tx (ShelleyLedgerEra era)

Convert a cardano-api Tx into a matching cardano-ledger Tx.

Re-export normal Tx (any era)

data Tx era