Hydra.Cardano.Api.Tx

Extras

txFee' :: HasCallStack => Tx era -> Lovelace

Get explicit fees allocated to a transaction.

NOTE: this function is partial and throws if given a Byron transaction for which fees are necessarily implicit.

totalExecutionCost :: PParams (ShelleyLedgerEra era) -> Tx era -> Lovelace

Calculate the total execution cost of a transaction, according to the budget assigned to each redeemer.

Type Conversions

toLedgerTx :: Tx Era -> ValidatedTx (ShelleyLedgerEra Era)

Convert a cardano-api Tx into a cardano-ledger Tx in the Alonzo era (a.k.a. ValidatedTx).

fromLedgerTx :: ValidatedTx (ShelleyLedgerEra Era) -> Tx Era

Convert a cardano-ledger's Tx in the Babbage era (a.k.a. ValidatedTx) into a cardano-api Tx.