Hydra.Cardano.Api.TxOut

Extras

txOuts' :: Tx era -> [TxOut CtxTx era]

modifyTxOutAddress :: (AddressInEra era -> AddressInEra era) -> TxOut ctx era -> TxOut ctx era

Alter the address of a TxOut with the given transformation.

modifyTxOutValue :: HasMultiAsset era => (Value -> Value) -> TxOut ctx era -> TxOut ctx era

Alter the value of a TxOut with the given transformation.

modifyTxOutDatum :: (TxOutDatum ctx0 era -> TxOutDatum ctx1 era) -> TxOut ctx0 era -> TxOut ctx1 era

Alter the datum of a TxOut with the given transformation.

findTxOutByAddress :: AddressInEra era -> Tx era -> Maybe (TxIn, TxOut CtxTx era)

Find first TxOut which pays to given address and also return the corresponding TxIn to reference it.

findTxOutByScript :: forall lang. HasPlutusScriptVersion lang => UTxO -> PlutusScript lang -> Maybe (TxIn, TxOut CtxUTxO Era)

Type Conversions

fromLedgerTxOut :: TxOut (ShelleyLedgerEra Era) -> TxOut ctx Era

Convert a cardano-ledger TxOut into a cardano-api TxOut

toLedgerTxOut :: TxOut CtxUTxO Era -> TxOut (ShelleyLedgerEra Era)

Convert a cardano-api TxOut into a cardano-ledger TxOut

fromPlutusTxOut :: (HasMultiAsset era, HasScriptData era, HasInlineDatums era, IsShelleyBasedEra era) => Network -> TxOut -> TxOut CtxUTxO era

Convert a plutus TxOut into a cardano-api TxOut. NOTE: Reference scripts are not resolvable right now. NOTE: Requires the Network discriminator (Testnet or Mainnet) because Plutus addresses are stripped off it.

toPlutusTxOut :: TxOut CtxUTxO Era -> Maybe TxOut

Convert a cardano-api TxOut into a plutus TxOut. Returns Nothing if a byron address is used in the given TxOut.